Rename git and svn classes
