+++ plugin/versioning/src/main/kotlin/net/twisterrob/gradle/vcs/VCSPluginExtension.kt
@@ -1,8 +1,8 @@
 package net.twisterrob.gradle.vcs
 
 import net.twisterrob.gradle.kotlin.dsl.extensions
-import net.twisterrob.gradle.vcs.git.GITPluginExtension
-import net.twisterrob.gradle.vcs.svn.SVNPluginExtension
+import net.twisterrob.gradle.vcs.git.GitPluginExtension
+import net.twisterrob.gradle.vcs.svn.SvnPluginExtension
 import org.gradle.api.Project
 import org.gradle.api.file.FileCollection
 import org.gradle.kotlin.dsl.getByName
@@ -13,10 +13,10 @@ abstract class VCSPluginExtension : VCSExtension {
 	var current: VCSExtension = DummyVcsExtension
 		internal set
 
-	val svn: SVNPluginExtension
-		get() = extensions.getByName<SVNPluginExtension>(SVNPluginExtension.NAME)
-	val git: GITPluginExtension
-		get() = extensions.getByName<GITPluginExtension>(GITPluginExtension.NAME)
+	val svn: SvnPluginExtension
+		get() = extensions.getByName<SvnPluginExtension>(SvnPluginExtension.NAME)
+	val git: GitPluginExtension
+		get() = extensions.getByName<GitPluginExtension>(GitPluginExtension.NAME)
 
 	// `: VCSExtension by current` is not possible because `current` is "lateinit"
 	override val isAvailable: Boolean
