+++ plugin/versioning/src/main/kotlin/net/twisterrob/gradle/vcs/git/GitPlugin.kt
@@ -1,4 +1,4 @@
-package net.twisterrob.gradle.vcs.svn
+package net.twisterrob.gradle.vcs.git
 
 import net.twisterrob.gradle.common.BasePlugin
 import net.twisterrob.gradle.kotlin.dsl.extensions
@@ -7,10 +7,10 @@ import org.gradle.api.Project
 import org.gradle.kotlin.dsl.create
 
 @Suppress("detekt.UnnecessaryAbstractClass") // Gradle convention.
-abstract class SVNPlugin : BasePlugin() {
+abstract class GitPlugin : BasePlugin() {
 
 	override fun apply(target: Project) {
 		super.apply(target)
-		project.vcs.extensions.create<SVNPluginExtension>(SVNPluginExtension.NAME, project.rootDir)
+		project.vcs.extensions.create<GitPluginExtension>(GitPluginExtension.NAME, project.rootDir)
 	}
 }
